新增feature: swal 组合键

+++ code/js.js
@@ -11,6 +11,7 @@ if ( 'OVER' != getCookie("FIRSTTIME") ) {
 };
 
 var urlcache = {};
+var currentSite;
 
 // 48 对应 0，90 对应 Z
 // 预处理，为已有设置键位添加网站avatar
@@ -23,31 +24,60 @@ for (var i = 48; i <= 90; i++) {
     }
 };
 
-//键位按下时 keydown
+// keydown
 $(document).keydown(function(ev) {
-    if(ev.ctrlKey) return false;
+    // if(ev.ctrlKey) return false;
+
+    // 组合键 ctrl + x
+    // 作用：设置键位
+    if(ev.ctrlKey) {
+        var code = String.fromCharCode(ev.keyCode);
+        $("#tempdate").val(code);
+        $("#LI_" + code).addClass("active");
+        // console.log('current code :' + code);
+        update();
+        return false;
+    }
+
+    // 组合键 alt + x
+    // 作用：删除键位
+    if(ev.altKey) {
+        var code = String.fromCharCode(ev.keyCode);
+        $("#tempdate").val(code);
+        $("#LI_" + code).addClass("active");
+        del();
+        return false;
+    }
+
+    // 回车确认跳转
+    if((ev.keyCode == 13) && currentSite){
+        console.log('13');
+        window.location.href = currentSite;
+        setTimeout(($("#message").html("正在跳转...")), 2000);
+        return false;
+    }
+
     var code = String.fromCharCode(ev.keyCode);
     $("#LI_" + code).addClass("active");
-    // console.log('down:' + "#LI_" + code);
     // setTimeout('$("#LI_' + code + '").removeClass("active");', 300);
     if(urlcache[code] == '' || typeof(urlcache[code]) == 'undefined') {
         console.log('down message')
-        $("#message").html('找不到这个按键的配置,注意切换您的输入法哦~~~');
-        setTimeout('$("#message").html("");', 2000)
+        $("#message").html('找不到这个按键的配置');
+        currentSite = null;
+        // setTimeout('$("#message").html("");', 2000)
     } else {
-        alert('keydown');
-        window.location.href = urlcache[code];
+        $("#message").html('你即将访问：' + urlcache[code] + '，回车确认');
+        currentSite = urlcache[code];
+        // window.location.href = urlcache[code];
     }
 });
 
-// 键位释放时 keyup
+// keyup
 $(document).keyup(function(ev) {
-    if(ev.ctrlKey) return false;
+    // if(ev.ctrlKey) return false;
     var code = String.fromCharCode(ev.keyCode);
-    // console.log(code);
     var key = "#LI_" + code;
     $(("#LI_" + code).toString()).removeClass("active");
-    // console.log("#LI_" + code);
 });
 
 // 鼠标移入时
@@ -72,31 +102,70 @@ $("#main li").mouseenter(function() {
 
 // 阻止事件传播
 function del() {
+
     var code = $("#tempdate").val();
-    urlcache[code] = '';
-    $("#" + code).remove();
-    deleteCookie("_" + code);
+    swal({   
+        title: "Are you sure?",   
+        text: "此操作将会清除该键位设置",   
+        type: "warning",   
+        showCancelButton: true,   
+        confirmButtonColor: "#DD6B55",   
+        confirmButtonText: "我确认",   
+        cancelButtonText: "反悔了",   
+        closeOnConfirm: false,   
+        closeOnCancel: false }, function(isConfirm){   
+        if (isConfirm) {     
+            swal("Deleted!", "该键位设置已重置", "success");   
+                
+                console.log('code :' + code);
+                urlcache[code] = '';
+                $("#" + code).remove();
+                deleteCookie("_" + code);
+        } else {     
+            swal("Cancelled", "你的键位设置依旧 :)", "error");   
+        } 
+    });
+ 
     return false;
 };
 
 // 阻止事件传播
 function update() {
     var code = $("#tempdate").val();
-    $("#LI_" + code).css('background', '#ccf');
-    var u = window.prompt("请输入键位 [" + code + "] 对应的网站地址", "");
-    $("#LI_" + code).css('background', '#fff');
-    if (u.indexOf('http://') == -1 && u.indexOf('https://') == -1) {
-        u = 'http://' + u
-    };
-    if (!IsURL(u)) {
-        alert('网站地址输入错误!请核对');
-        return false;
-    };
-    urlcache[code] = u;
-    $("#" + code).remove();
-    $("#LI_" + code).prepend('<img id="' + code + '" class="fav" src="' + getico(u) + '" align="center">');
-    setCookie("_" + code, u);
-    return true;
+    // $("#LI_" + code).css('background', '#ccf');
+    // var u = window.prompt("请输入键位 [" + code + "] 对应的网站地址", "");
+    var u;
+
+    swal({   
+        title: "Hello",   
+        text: "请输入键位 [" + code + "] 对应的网站地址",   
+        type: "input",   
+        showCancelButton: true,   
+        closeOnConfirm: false,   
+        animation: "slide-from-top",   
+        inputPlaceholder: urlcache[code] }, function(u){   
+            if (u === false) return false;      
+            if (u === "") {     
+                swal.showInputError("You need to write something!");     
+                return false   
+            }
+            if (u.indexOf('http://') == -1 && u.indexOf('https://') == -1) {
+                u = 'http://' + u
+            };
+            if (!IsURL(u)) {
+                // alert('网站地址输入错误!请核对');
+                // console.log('wrong');
+                swal.showInputError("网址地址输入有误，请核对！");     
+                return false;
+            };      
+        swal("Nice!", "You wrote: " + u, "success");
+        // $("#LI_" + code).css('background', '#fff');
+        urlcache[code] = u;
+        // $("#" + code).remove();
+        $("#LI_" + code).prepend('<img id="' + code + '" class="fav" src="' + getico(u) + '" align="center">');
+        setCookie("_" + code, u);
+        return true;
+    });    
 };
 
 function getico(url) {
